+++ src/index.ts
@@ -192,4 +192,4 @@ export {
 export { iteratePages, collectAllPages } from './utils/index.js';
 export type { WithoutUndefined } from './utils/index.js';
 
-export const VERSION = '0.3.0';
+export const VERSION = '0.3.1';